requirements.txt
openai>=1.3.7
httpx>=0.26.0
tiktoken>=0.5.1
pydantic>=2.5.3
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
